% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{build_model}
\alias{build_model}
\title{Building model}
\usage{
build_model(training_set, recipe, optimal, vfold, gridvals, k)
}
\arguments{
\item{training_set}{dataframe containing the data to train the model on}

\item{recipe}{a recipe to prepare the data for modelling}

\item{optimal}{a variable that takes in the string "None" or "Yes" to specify whether
to use tune() for the neighbors("None" optimal yet) or use an existing number of optimal k}

\item{vfold}{dataframe containing data splitted into folds for cross validation}

\item{gridvals}{tibble containing values of k for tuning}

\item{k}{Additional numeric argument, take in an integer to indicate the number of
optimal neighbors for the model}
}
\value{
a fitted knn model

' @example build_model(image_recipe, NULL)
}
\description{
Function build_model() to build knn model with cross validation
}
