% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Create a tidied data frame}
\usage{
load_data(path, names = TRUE, separator = ",", na_values = "NAs")
}
\arguments{
\item{path}{A single string or a vector of the path to raw file}

\item{names}{A vector of names of the columns}

\item{separator}{A string of the separator used in the file}

\item{na_values}{A string of how the NAs values are represented in the file}
}
\value{
a new tidied data frame with appropriate columns type
}
\description{
Reading in a data frame and change column types from characters to
factors
}
\examples{
  load_data(readr::readr_example("mini-gapminder-asia.csv"))

}
