% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist.R
\name{plot_hist}
\alias{plot_hist}
\title{Abstraction of histograms for factorial variables}
\usage{
plot_hist(data, class_column, sep = ".", bins = 20, col = 2, title)
}
\arguments{
\item{data}{the data frame}

\item{class_column}{the groups that we are going to grouped the variables by}

\item{sep}{the character that is seperating the column names}

\item{bins}{a number that indicates
the number of bins in histograms}

\item{col}{the number of columns when arranging the plots}

\item{title}{a string that indicates the plot title}
}
\value{
A histogram of all numeric variables that is
grouped by the class_Column

' @example
plot_hist(iris, Species, title = "Histogram")
plot_hist(iris, Species, binwidth = 0.25, col = 3, title = "Histogram")
}
\description{
Create a general function to remove repetitive ggplot functions
}
