% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_numeric.R
\name{avg_numeric}
\alias{avg_numeric}
\title{Compute averages of numeric attributes}
\usage{
avg_numeric(dataset, class_col)
}
\arguments{
\item{dataset}{the training set with a list of numerical attributes}

\item{class_col}{the name for the column to group by

'@example avg_numeric(iris,Species)}
}
\description{
Function to compare the mean values of the numerical attributes
}
