% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Create a tidied data frame}
\usage{
load_data(path, names, separator = ",", na_values = "NAs")
}
\arguments{
\item{path}{A single string or a vector of the path to raw file}

\item{names}{A vector of names of the columns}

\item{separator}{A string of the separator used in the file}

\item{na_values}{A string of how the NAs values are represented in the file}
}
\value{
a new tidied data frame with appropriate columns type

'@examples
load_data("data/raw/heart_disease_data.csv",
names = c('age','sex','chest_pain_type','resting_bp','cholesterol',
'high_blood_sugar','resting_ecg','max_heart_rate',
'exercise_pain','old_peak','slope','no_vessels_colored',
'thal','diagnosis','diagnosis_2'),
separator = ",",
na_values = "?")

load_data("data/raw/heart_disease_data.csv",
c('age','sex','chest_pain_type','resting_bp','cholesterol',
'high_blood_sugar','resting_ecg',
'max_heart_rate','exercise_pain','old_peak','slope',
'no_vessels_colored','thal',
'diagnosis','diagnosis_2'))
}
\description{
Reading in a data frame and change column types from characters to
factors
}
